backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Vector Database & Embeddings
qdrant-client==1.7.3
sentence-transformers==2.3.1
torch>=2.0.0

# Text Processing & NLP
openai==1.10.0
groq>=0.4.0
tiktoken==0.5.2
nltk==3.8.1

# Document Processing
pypdf2==3.0.1
python-multipart==0.0.6
# HTML / File utilities
beautifulsoup4==4.12.3
python-docx==1.1.2
# python-magic==0.4.27  # Commented out - can cause issues on macOS

# Graph & Relationships
networkx==3.2.1

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1

